requirements.txt
streamlit==1.38.0
transformers==4.44.2
torch>=2.2.0
sentencepiece>=0.1.99
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
